hi-buddy-all/20multimodal_optimize_pyav_fer/requirements.txt
fastapi
uvicorn[standard]
aiortc
av
numpy
soundfile
resampy
onnxruntime
pyttsx3
gTTS
torch
torchvision
transformers
Pillow
